Capstone-project/backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
requests
python-dotenv
google-generativeai
python-multipart
aiofiles
gtts
numpy>=1.26.4
sqlalchemy>=2.0
psycopg2-binary
PyJWT
librosa>=0.10
soundfile
# Voice cloning — install manually: pip install TTS
# TTS  ← uncomment after: pip install TTS
